backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
rq==1.15.1
boto3==1.34.14
openai==1.9.0
pdf2image==1.17.0
PyPDF2==3.0.1
PyMuPDF==1.23.26
pillow==10.2.0
opencv-python==4.9.0.80
pytesseract==0.3.10
openpyxl==3.1.2
pandas==2.1.4
numpy==1.26.3
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
python-multipart==0.0.6
aiofiles==23.2.1
Jinja2==3.1.3
psutil==5.9.8